Use symbolicmode.chmod to simplify chmod() code.
symbolicmode provides a chmod() method that can take symbolic or numeric
permissions, so the function can be simplified.

Note, this depends on symbolicmode 2.x, because I decided (as I was
reviewing the use here) that I had the order of the chmod arguments
backwards, so I made the backwards-incompatible change and rolled
the major version up because of that.